Try to set proper genesis parachain inherent.
