+++ wardrobe/parachain/Cargo.toml
@@ -22,15 +22,15 @@ sp-std = { default_features = false, workspace = true }
 # Consider dropping this one.
 tuxedo-core = { default-features = false, path = "../../tuxedo-core" }
 tuxedo-parachain-core = { default-features = false, path = "../../tuxedo-parachain-core" }
-
-[dev-dependencies]
-cumulus-primitives-core = { workspace = true }
-cumulus-test-relay-sproof-builder = { workspace = true }
+cumulus-primitives-core = { workspace = true, optional = true }
+cumulus-test-relay-sproof-builder = { workspace = true, optional = true }
 
 [features]
 default = [ "std" ]
 std = [
 	"cumulus-primitives-parachain-inherent/std",
+	"cumulus-primitives-core",
+	"cumulus-test-relay-sproof-builder",
 	"log/std",
 	"parity-scale-codec/std",
 	"scale-info/std",
